State Manager moved outside of state class
